requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.28.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
jsonpickle==3.0.2
kiwisolver==1.4.5
magic-filter==1.0.12
matplotlib==3.8.0
multidict==6.0.4
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.7.1
yarl==1.9.2
